requirements.txt
Django==2.2
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
Pillow==7.2.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
